backend/requirements/base.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic[email]
pydantic-settings
supabase
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-generativeai
langchain
langchain-google-genai
sentence-transformers
chromadb
PyPDF2
pypdf
python-docx
mammoth
tiktoken
numpy
pandas
scikit-learn
loguru
groq
mem0ai
google-genai